move config files around
