+++ configs/countryConfigs/equatorial-guinea.js
