+++ configs/countryConfigs/gabon-parks.js
