+++ configs/countryConfigs/paraguay.js
