--- configs/leftPanel.config.ts
@@ -1,258 +1,0 @@
-const drawTool = 'DRAW_ICON_HERE';
-
-export const analysisContent = {
-  en: {
-    analyzeExistingShapeTitle: 'Analyze a shape on the map',
-    analyzeExistingShapeDirections: [
-      'Use the layers tab to turn on a data layer',
-      'Select a shape on the map',
-      'Click on the analyze tab'
-    ],
-    analyzeYourShapeTitle: 'Analyze your own shape',
-    analyzeYourShapeDirections: [
-      `Choose the draw tool ${drawTool} in the toolbox`,
-      'Draw a shape anywhere on the map',
-      'Select the shape to run the analysis'
-    ],
-    drawButton: 'Start drawing',
-    enterCoordinatesTitle: 'Enter your own coordinates',
-    enterCoordinatesDirections: [
-      `Enter at least 3 coordinates`,
-      `Add up to 10 points to make shapes`
-    ],
-    coordinatesButton: 'Enter Values',
-    visitTitle: 'Add additional shapes in the future by visiting the draw tool'
-  },
-  ka: {
-    analyzeExistingShapeTitle: 'ჩაატარეთ ფიგურის ანალიზი რუკაზე',
-    analyzeExistingShapeDirections: [
-      'მონაცემების ფენის ჩასართავად გამოიყენეთ ფენების მენიუ',
-      'აარჩიეთ ფიგურა რუკაზე',
-      'დააჭირეთ ანალიზისი მენიუს'
-    ],
-    analyzeYourShapeTitle: 'ჩაატარეთ თქვენი ფიგურის ანალიზი',
-    analyzeYourShapeDirections: [
-      `შეარჩიეთ სახაზავი ინსტრუმენტი ${drawTool} ტულბოქსიდან`,
-      'რუკაზე ნებისმიერი ფიგურა დახაზეთ',
-      'ანალიზის ჩასატარებლად შეარჩიეთ ფიგურა'
-    ],
-    drawButton: 'დაიწყეთ ხაზვა',
-    enterCoordinatesTitle: 'შეიტანეთ თქვენი კოორდინატები',
-    enterCoordinatesDirections: [
-      `შეიტანეთ მინიმუმ  3 კოორდინატი`,
-      `დაამატეთ 10 წერტილამდე პოლიგონის შესაქმნელად`
-    ],
-    coordinatesButton: `მნიშვნელობების შეტანა`,
-    visitTitle: `მომავალში დაამატეთ დამატებითი ფიგურები სახაზავი ინსტრუმენტის ${drawTool} მეშვეობით`
-  },
-  fr: {
-    analyzeExistingShapeTitle: 'Analysez un polygone sur la carte',
-    analyzeExistingShapeDirections: [
-      "Utilisez l'onglet Couches pour activer une couche de données",
-      'Sélectionnez une entité sur la carte',
-      "Cliquez sur l'onglet analyse"
-    ],
-    analyzeYourShapeTitle: 'Analysez votre propre polygone',
-    analyzeYourShapeDirections: [
-      `Utiliser l'outil dessin ${drawTool} dans la boîte à outil`,
-      'Tracez un polygone sur la carte',
-      "Cliquez sur le polygone pour lancer l'analyse"
-    ],
-    drawButton: 'Débutez le dessin',
-    enterCoordinatesTitle: 'Entrez vos propres coordonnées',
-    enterCoordinatesDirections: [
-      `Entrez au moins 3 coordonnées`,
-      `Ajoutez jusqu’à 10 points pour définir votre zone`
-    ],
-    coordinatesButton: `Entrez les valeurs`,
-    visitTitle: "Ajouter des polygones supplémentaires grâce à l'outil dessin"
-  },
-  es: {
-    analyzeExistingShapeTitle: 'Analizar un polígono  en el mapa',
-    analyzeExistingShapeDirections: [
-      'Usar la pestaña de capas para encender una capa',
-      'Seleccionar un polígono en el mapa',
-      'Hacer clic en la pestaña de analizar'
-    ],
-    analyzeYourShapeTitle: 'Analizar su propio polígono',
-    analyzeYourShapeDirections: [
-      `Escoger la herramienta de dibujo ${drawTool} en la caja de herramientas`,
-      'Dibujar un polígono en cualquier lugar del mapa',
-      'Seleccionar el polígono para hacer el análisis'
-    ],
-    drawButton: 'Empezar a dibujar',
-    enterCoordinatesTitle: 'Ingresa las coordenadas',
-    enterCoordinatesDirections: [
-      `Ingresa por lo menos 3 coordenadas`,
-      `Agrega hasta 10 puntos para generar el polígono`
-    ],
-    coordinatesButton: `Ingresa valores`,
-    visitTitle:
-      'Añadir polígonos adicionales en el futuro usando la herramienta de dibujo '
-  },
-  pt: {
-    analyzeExistingShapeTitle: 'Analise a área selecionada no mapa',
-    analyzeExistingShapeDirections: [
-      'Use a guia camadas para ativar os dados da camada',
-      'Selecione uma área no mapa',
-      'Clique na guia de Análise'
-    ],
-    analyzeYourShapeTitle: 'Use sua propria área de Análise',
-    analyzeYourShapeDirections: [
-      `Escolha a ferramenta de desenho ${drawTool} na caixa de ferramentas`,
-      'Desenhe uma área de interesse no mapa',
-      'Selecione a área na qual irá executar a análise'
-    ],
-    drawButton: 'Comece a desenhar',
-    enterCoordinatesTitle: 'Entre com suas coordenadas',
-    enterCoordinatesDirections: [
-      `Insira pelo menos 3 coordenadas`,
-      `Adicione até 10 pontos para desenhar o shape`
-    ],
-    coordinatesButton: `Entre com os valores`,
-    visitTitle:
-      'Adicionar formas adicionais no futuro, visitando a ferramenta de desenho'
-  },
-  id: {
-    analyzeExistingShapeTitle: 'Analyze a shape on the map',
-    analyzeExistingShapeDirections: [
-      'Use the layers tab to turn on a data layer',
-      'Select a shape on the map',
-      'Click on the analyze tab'
-    ],
-    analyzeYourShapeTitle: 'Analyze your own shape',
-    analyzeYourShapeDirections: [
-      `Choose the draw tool ${drawTool} in the toolbox`,
-      'Draw a shape anywhere on the map',
-      'Select the shape to run the analysis'
-    ],
-    drawButton: 'Start drawing',
-    enterCoordinatesTitle: 'Masukkan koordinat Anda',
-    enterCoordinatesDirections: [
-      `Masukkan paling sedikit 3 koordinat`,
-      `Tambahkan sampai 10 titik untuk membuat bentuk`
-    ],
-    coordinatesButton: `Masukkan Nilai`,
-    visitTitle: 'Add additional shapes in the future by visiting the draw tool '
-  },
-  zh: {
-    analyzeExistingShapeTitle: '分析地图上的图形',
-    analyzeExistingShapeDirections: [
-      '使用图层键来打开图层',
-      '在地图上选择一个图形',
-      '点击分析键'
-    ],
-    analyzeYourShapeTitle: '',
-    analyzeYourShapeDirections: [
-      `在工具箱里选择一个绘画工具 ${drawTool}`,
-      '在地图上任意地方画一个图形',
-      '选择进行分析的图形'
-    ],
-    drawButton: '开始绘画',
-    enterCoordinatesTitle: '输入你的坐标',
-    enterCoordinatesDirections: [
-      `输入至少三个坐标`,
-      `添加最多10个坐标来完成作图`
-    ],
-    coordinatesButton: `输入`,
-    visitTitle: '使用绘画工具来添加更多图形'
-  }
-};
-
-export const basemapLayersContent = {
-  defaultESRIBasemaps: [
-    {
-      id: 'satellite',
-      thumbnailUrl:
-        'https://my.gfw-mapbuilder.org/js/arcgis-api-mapbuilder-1.2/esri/images/basemap/satellite.jpg',
-      title: {
-        en: 'Imagery',
-        ka: 'Imagery',
-        fr: 'Imagery',
-        es: 'Imagery',
-        pt: 'Imagery',
-        id: 'Imagery',
-        zh: 'Imagery'
-      }
-    },
-    {
-      id: 'hybrid',
-      thumbnailUrl:
-        'https://my.gfw-mapbuilder.org/js/arcgis-api-mapbuilder-1.2/esri/images/basemap/hybrid.jpg',
-      title: {
-        en: 'Imagery with Labels',
-        ka: 'Imagery with Labels',
-        fr: 'Imagery with Labels',
-        es: 'Imagery with Labels',
-        pt: 'Imagery with Labels',
-        id: 'Imagery with Labels',
-        zh: 'Imagery with Labels'
-      }
-    },
-    {
-      id: 'osm',
-      thumbnailUrl:
-        'https://my.gfw-mapbuilder.org/js/arcgis-api-mapbuilder-1.2/esri/images/basemap/osm.jpg',
-      title: {
-        en: 'Open Street Map',
-        ka: 'Open Street Map',
-        fr: 'Open Street Map',
-        es: 'Open Street Map',
-        pt: 'Open Street Map',
-        id: 'Open Street Map',
-        zh: 'Open Street Map'
-      }
-    }
-  ]
-};
-
-export const documentsContent = {
-  en: {
-    instructions:
-      'Select an area of interest to see if there are any related documents',
-    name: 'Name',
-    pdf: 'PDF',
-    size: 'Size'
-  },
-  ka: {
-    instructions:
-      'შეარჩიეთ საინტერესო ტერიტორია, და ნახეთ არის თუ არა მასთან დაკავშირებული დოკუმენტები',
-    name: 'სახელი',
-    pdf: 'PDF',
-    size: 'ზომა'
-  },
-  fr: {
-    instructions:
-      "Choisisez une région d'intérêt pour voir les documents associés",
-    name: 'Nom du fichier',
-    pdf: 'PDF',
-    size: 'Taille'
-  },
-  es: {
-    instructions:
-      'Seleccionar una área de interés para ver si haya algún documento relacionado',
-    name: 'Nombre',
-    pdf: 'PDF',
-    size: 'Size'
-  },
-  pt: {
-    instructions:
-      'Selecione área de interesse para verificar a existência de documentos relacionados',
-    name: 'Nome',
-    pdf: 'PDF',
-    size: 'Tamanho'
-  },
-  id: {
-    instructions:
-      'Select an area of interest to see if there are any related documents',
-    name: 'Nama',
-    pdf: 'PDF',
-    size: 'Size'
-  },
-  zh: {
-    instructions: '选择感兴趣区域来查看相关文件',
-    name: '姓名',
-    pdf: 'PDF',
-    size: 'Size'
-  }
-};
